requirements.txt
streamlit
openai
google-genai
Pillow